+++ internal/match.go
@@ -5,7 +5,7 @@ type Match struct {
 	Strategy strategy
 }
 
-// MatchStrings returns float32
-func (m *Match) MatchStrings(s1 string, s2 string) uint16 {
+// MatchStrings returns int
+func (m *Match) MatchStrings(s1 string, s2 string) int {
 	return m.Strategy.matchStrings(s1, s2)
 }
